requirements.txt
pynput>=1.7.6
sounddevice>=0.4.6
scipy>=1.10.0
numpy>=1.24.0
openai>=1.0.0
pyperclip>=1.8.2
pystray>=0.19.0
Pillow>=10.0.0
python-dotenv>=1.0.0
